+++ src/SpinoramaSlideshow/SpinoramaSlideshow.types.ts
@@ -1,3 +1,1 @@
-/// <reference types="react" />
-
 export type SpinoramaSlideshowProps = { className?: string }