+++ src/SpinoramaSlideshow/index.ts
@@ -1,2 +1,3 @@
 export { SpinoramaSlideshow } from './SpinoramaSlideshow'
-export * from './SpinoramaSlideshow'
+
+export * from './SpinoramaSlideshow.types'