+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this package will be documented in this file.
 
 - Change `README.md` for **under development**
+- Export `types` from all files
 
 # 0.1.3
 - Create `SpinoramaSlideshow` component